Java impl
