+++ src/test/java/iog/psg/cardano/jpi/JpiResponseCheck.java
@@ -15,7 +15,7 @@ import scala.util.Either;
 
 public class JpiResponseCheck {
 
-    public final CardanoApi jpi;
+    public final CardanoApiImpl jpi;
     private final long timeout;
     private final TimeUnit timeoutUnit;
 
@@ -25,7 +25,7 @@ public class JpiResponseCheck {
         timeoutUnit = null;
     }
 
-    public JpiResponseCheck(CardanoApi jpi, long timeout, TimeUnit timeoutUnit) {
+    public JpiResponseCheck(CardanoApiImpl jpi, long timeout, TimeUnit timeoutUnit) {
         this.jpi = jpi;
         this.timeoutUnit = timeoutUnit;
         this.timeout = timeout;
@@ -109,7 +109,7 @@ public class JpiResponseCheck {
         return jpi.getTransaction(walletId, txId).toCompletableFuture().get(timeout, timeoutUnit);
     }
 
-    public static CardanoApi buildWithPredefinedApiExecutor(iog.psg.cardano.ApiRequestExecutor executor, ActorSystem as) {
+    public static CardanoApiImpl buildWithPredefinedApiExecutor(iog.psg.cardano.ApiRequestExecutor executor, ActorSystem as) {
         CardanoApiBuilder builder = CardanoApiBuilder.create("http://fake:1234/").withApiExecutor(new ApiRequestExecutor() {
             @Override
             public <T> CompletionStage<T> execute(iog.psg.cardano.CardanoApiImpl.CardanoApiRequest<T> request) throws CardanoApiException {
@@ -122,7 +122,7 @@ public class JpiResponseCheck {
         return builder.build();
     }
 
-    public static CardanoApi buildWithDummyApiExecutor() {
+    public static CardanoApiImpl buildWithDummyApiExecutor() {
         CardanoApiBuilder builder = CardanoApiBuilder.create("http://fake/").withApiExecutor(new ApiRequestExecutor() {
             @Override
             public <T> CompletionStage<T> execute(iog.psg.cardano.CardanoApiImpl.CardanoApiRequest<T> request) throws CardanoApiException {
