+++ src/test/scala/iog/psg/cardano/CardanoJpiSpec.scala
@@ -4,7 +4,7 @@ import java.time.ZonedDateTime
 import java.util.concurrent.CompletionStage
 
 import akka.actor.ActorSystem
-import iog.psg.cardano.jpi.{AddressFilter, JpiResponseCheck, ListTransactionsParamBuilder}
+import iog.psg.cardano.jpi.{AddressFilter, CardanoApiImpl, JpiResponseCheck, ListTransactionsParamBuilder}
 import iog.psg.cardano.util._
 import org.scalatest.Assertions
 import org.scalatest.concurrent.ScalaFutures
@@ -211,7 +211,7 @@ class CardanoJpiSpec
     with DummyModel
     with JsonFiles {
     override implicit val as: ActorSystem = getCurrentSpecAS
-    lazy val customApi: jpi.CardanoApi = JpiResponseCheck.buildWithPredefinedApiExecutor(inMemoryExecutor, as)
+    lazy val customApi: CardanoApiImpl = JpiResponseCheck.buildWithPredefinedApiExecutor(inMemoryExecutor, as)
   }
 
 }
